more effects repaired

+++ src/common/CSpriteObject.cpp
@@ -407,8 +407,8 @@ void CSpriteObject::draw()
 
 	if( scrx < gameres.w && scry < gameres.h && exists )
 	{
-		Uint16 showX = scrx+Sprite.getXOffset();
-		Uint16 showY = scry+Sprite.getYOffset();
+        int showX = scrx+Sprite.getXOffset();
+        int showY = scry+Sprite.getYOffset();
 		if(m_blinktime > 0)
 		{
 			Sprite.drawBlinkingSprite( showX, showY );
