+++ src/engine/galaxy/common/ai/CStunnable.cpp
@@ -116,8 +116,8 @@ void CStunnable::draw()
       
       if( scrx < gameres.w && scry < gameres.h && exists )
       {
-        Uint16 showX = scrx+StarSprite.getXOffset();
-        Uint16 showY = scry+StarSprite.getYOffset();
+        int showX = scrx+StarSprite.getXOffset();
+        int showY = scry+StarSprite.getYOffset();
         StarSprite.drawSprite( showX, showY, (255-transluceny) );
       }
     }
